refactor: move CostString generation to a class for reuse
